Refactor ContentStoreServer: Split into multiple modules for better organization
- Moved tool argument types to a new `types.rs` file for clarity.
- Created a `schemas.rs` file to define tool schemas separately.
- Extracted server implementation into `server.rs` for improved structure.
- Removed redundant code and improved modularity by re-exporting necessary components.
- Updated `mod.rs` to reflect new module structure and maintain functionality.
